update tlms.rs to latest master, rework correlate to support db
