+++ Cargo.toml
@@ -4,15 +4,14 @@ version = "0.2.0"
 edition = "2021"
 
 [dependencies]
-tlms = { version = "0.8", git = "https://github.com/tlm-solutions/tlms.rs", tag = "v0.8.0", features = ["schema", "measurements", "telegrams", "trekkie", "management"]}
-lofi = { version = "0.3", git = "https://github.com/tlm-solutions/lofi", tag = "v0.3.1", features = [ "correlate" ] }
+tlms = { version = "0.9", git = "https://github.com/tlm-solutions/tlms.rs", features = [ "measurements", "telegrams", "trekkie", "management"]}
+lofi = { version = "0.3", git = "https://github.com/tlm-solutions/lofi", features = [ "correlate" ] }
 
 clap = { version = "4.1", features = ["derive"] }
 
 # webserver shit
 actix = "0.13"
 actix-web = "4.0"
-actix-diesel = { git = "https://github.com/frol/actix-diesel", branch = "actix-0.13", features = ["actix-web"] }
 actix-identity = "*"
 actix-session = { version = "0.7", features = ["redis-actor-session"] }
 actix-multipart = "*"
